Only inter-group bonds in main bondAtomList field
The specification does not state that only the inter group bonds are expected in the main bondAtomList field, except implicitly in the pseudo-code for decoding. If one writes an encoder that writes all bonds in the main bondAtomList, the output mmtf file would break many parsers implementations and the file would have poor compression ratios.